passing component to a file with its name
